requirements.txt
absl-py==2.2.2
addict==2.4.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chumpy==0.70
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.12
einops==0.8.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
json-tricks==3.17.3
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==1.3.9
mpmath==1.3.0
munkres==1.1.4
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.8.0.87
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
packaging==24.2
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scipy==1.10.1
-e git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf#egg=segment_anything
six==1.17.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
timm==0.3.2
tomli==2.2.1
torch==1.13.1+cu116
torchaudio==0.13.1+rocm5.2
torchvision==0.14.1+cu116
tqdm==4.67.1
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1727769808118/work
Werkzeug==3.0.6
xtcocotools==1.14.3
yapf==0.43.0
zipp==3.20.2
